% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unfolding.R
\name{binning3d}
\alias{binning3d}
\title{Histogram data}
\usage{
binning3d(size, angle, shape, breaks, check = TRUE, na.rm = TRUE)
}
\arguments{
\item{size}{vector of sizes}

\item{angle}{vector of angles}

\item{shape}{vector of shape factors}

\item{breaks}{list of bin vectors}

\item{check}{logical, default is \code{TRUE}}

\item{na.rm}{logical, whether \code{NAs} should be removed, default \code{TRUE}}
}
\value{
3D array of binned count data
}
\description{
Count data of size, shape and orientation
}
\details{
For each value of planar or spatial measured quantities \code{size}, \code{shape} and \code{orientation}
the function counts the number of observations falling into each class (bin). The list \code{breaks} can be
obtained by the function \code{\link{setbreaks}}. If \code{check=TRUE}, some checks on \code{breaks} are done.
For an example please see the file 'almmc.R'.
}
\author{
M. Baaske
}
