% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cylinder.R
\name{cylinders3d}
\alias{cylinders3d}
\title{3D a cylinder system}
\usage{
cylinders3d(S, box, draw.axes = FALSE, draw.box = TRUE, clipping = FALSE,
  ...)
}
\arguments{
\item{S}{a list of cylinders}

\item{box}{simulation box}

\item{draw.axes}{logical, if \code{TRUE}, draw the axes}

\item{draw.box}{logical, if \code{TRUE}, draw the bounding box}

\item{clipping}{logical, if \code{TRUE}, clip to the bounding box}

\item{...}{further material properties passed to 3d plotting functions}
}
\description{
Draw 3D spherocylinders
}
\details{
The function requires the package \code{rgl} to be installed.
}
\author{
M. Baaske
}
