% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/spheroid.R
\name{getAngle}
\alias{getAngle}
\title{Map angle}
\usage{
getAngle(phi)
}
\arguments{
\item{phi}{angle}
}
\value{
transformed angle
}
\description{
Transform orientation angle
}
\details{
The function transforms the colatitude \eqn{\theta\in[0,2\pi]}, the angle between
the rotational symmetry axis of the spheroid and the vertical axis of the simulation box
to \eqn{[0,\pi/2]}.
}

