% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sphere.R
\name{getSphereSystem}
\alias{getSphereSystem}
\title{Get sphere system}
\usage{
getSphereSystem(S)
}
\arguments{
\item{S}{result of \code{\link{simSphereSystem}}}
}
\value{
list of spheres of class \code{spheres}
}
\description{
Get the stored sphere system
}
\details{
The sphere system is internally stored as a C data structure.
This function copies and converts the sphere objects to the R level.
}
\author{
M. Baaske
}
