% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spheroid.R
\name{intersectSystem}
\alias{intersectSystem}
\title{Intersect a 3D system of grains}
\usage{
intersectSystem(S, d, n = c(0, 1, 0), intern = FALSE, pl = 0)
}
\arguments{
\item{S}{list of spheroids, see \code{\link{simPoissonSystem}}}

\item{d}{distance of the the box-aligned intersecting plane from the origin}

\item{n}{normal vector which defines the intersting plane}

\item{intern}{logical, \code{FALSE} (default), return all section profiles otherwise
only those which have their centers inside the intersection window (if the 
intersected spheroid system had been simulated using exact simulation)}

\item{pl}{integer, \code{pl=0} (default), only return pointer to stored intersections}
}
\value{
list of size, shape and angle of section profiles
}
\description{
Intersect a system of spheres, spheroids or cylinders by a given plane
}
\details{
The function intersects a given system of spheres, spheroids or cylinders by a plane which is defined by a
normal vector, e.g. \code{n=c(0,1,0)}, perpendicular to one of the lateral planes of the simulation box. 
The print level \code{pl>=0} sets the type of return value. If \code{pl=10} the functions only returns
the lengths of the semi axes, the shape factor of the angle in the intersection plane between \code{[0,pi/2]}.
}
\author{
M. Baaske
}
