% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unfolding.R
\name{parameterEstimates}
\alias{parameterEstimates}
\title{Spatial histogram}
\usage{
parameterEstimates(H, breaks)
}
\arguments{
\item{H}{trivariate (output) histogram, estimated by \code{\link{unfold}}}

\item{breaks}{breaks vector from \code{\link{setbreaks}}}
}
\value{
list of the following entries: either major or minor semi-axis length \code{a},
        (polar) angle \code{Theta} and shape factor \code{s}, see \code{\link{parameters3d}}
}
\description{
Characteristics for a spatial histogram
}
\details{
Based on the estimated joint distribution after unfolding the function replicates
the entries of the vector \code{breaks}, see \code{\link{setbreaks}}, equal to the
number of estimated count data for each class. For an example please see the file 'unfold.R'.
}
\author{
M. Baaske
}
