% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unfolding.R
\name{parameters3d}
\alias{parameters3d}
\title{Get 3D characteristics of spheroids}
\usage{
parameters3d(S)
}
\arguments{
\item{S}{list of spheroids}
}
\value{
a list of spheroids` 3D characteristics
}
\description{
Extract the original 3D size-shape-orientation characteristics
}
\details{
The function simply extracts the characteristics of a 3D spheroid system
either as oblate or prolate type and returns a list which consists of entries
named 'a' (major semi-axis length), 's' (shape factor) and polar angle 'Theta'.
For an example please see the file 'unfold.R'.
}
\author{
M. Baaske
}
