% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sphere.R
\name{setupSphereSystem}
\alias{setupSphereSystem}
\title{Setup sphere system}
\usage{
setupSphereSystem(S, box = list(c(0, 1)), perfect = TRUE, pl = 0,
  label = "N")
}
\arguments{
\item{S}{sphere system}

\item{box}{simualtion box}

\item{perfect}{logical, \code{TRUE} (default), simulate perfect}

\item{pl}{print level}

\item{label}{some character as a label, `\code{N}` (default)}
}
\value{
\code{NULL}
}
\description{
Reinitialize sphere system after R workspace reloading
}
\details{
The internally stored sphere system has to be reinitialized
after R workspace reloading. Calling this function is needed only in case one desires to get profile
sections of the sphere system again which has been previously stored as an R (list) object and afterwards
reloaded.
}
\author{
M. Baaske
}
