% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/spheroid.R
\name{setupSpheroidSystem}
\alias{setupSpheroidSystem}
\title{Setup spheroid system}
\usage{
setupSpheroidSystem(S, box, pl = 0)
}
\arguments{
\item{S}{spheroid system}

\item{box}{simulation box}

\item{pl}{print level}
}
\value{
\code{NULL}
}
\description{
Reinitialize spheroid system after workspace reloading
}
\details{
The internally stored spheroid system has to be reinitialized
 after R workspace reloading. Calling this function is needed only in case one desires to get different
 section profiles of the spheroid system which has been stored as an R (list) object first and reloaded in
 the user workspace again.
}
\seealso{
\code{\link{simSpheroidSystem}}
}

