% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cylinder.R
\name{simCylinderIntersection}
\alias{simCylinderIntersection}
\title{Cylinder intersection}
\usage{
simCylinderIntersection(theta, cond)
}
\arguments{
\item{theta}{simulation parameters}

\item{cond}{conditioning object for simulation and intersection}
}
\value{
list of intersection profiles
}
\description{
Simulate a cylinder system and intersect
}
\details{
The function first simulates a cylinder system according to the parameter \code{theta} and only
returns the resulting section profiles.
}
