% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cylinder.R
\name{simCylinderSystem}
\alias{simCylinderSystem}
\title{Simulation of cylinder system}
\usage{
simCylinderSystem(theta, lam, size = "const", shape = "const",
  orientation = "rbetaiso", rjoint = NULL, box = list(c(0, 1)),
  mu = c(0, 1, 0), dz = 0, n = c(0, 1, 0), profiles = FALSE,
  intern = FALSE, perfect = TRUE, pl = 0, label = "N")
}
\arguments{
\item{theta}{simulation parameters (see examples)}

\item{lam}{mean number of spheroids per unit volume}

\item{size}{\code{size="const"} (default) or name of random generating function
a for specific size distribution}

\item{shape}{either \code{shape="const"} as a constant portion of the axis length
   or \code{shape="rbeta"} as beta distributed shape factor. For a radius distribution
use your own joint distribution, see details.}

\item{orientation}{name of random generating function for orientation distribution}

\item{rjoint}{name of joint random generating function}

\item{box}{simulation box}

\item{mu}{main orientation axis, \code{mu=c(0,1,0)} (default)}

\item{dz}{distance of the intersecting x-, y-plane to the origin}

\item{n}{normal vector of intersting plane}

\item{profiles}{logical, \code{profiles=FALSE} (default), whether the simulated system of cylinder is intersected afterwards in which case
only sections profiles are returned}

\item{intern}{logical, \code{intern=FALSE} (default), whether to return only section profiles with centers inside the simulation window}

\item{perfect}{logical, \code{perfect=TRUE} (default) simulate perfect}

\item{pl}{integer, print level and return value definition}

\item{label}{character, label passed to each simulated cylinder, `\code{N}` (default)}
}
\value{
list of cylinders
}
\description{
Simulation of Poisson cylinder system
}
\details{
The function simulates a Poisson (sphero)cylinder system according to the supplied
simulation parameter \code{theta} in a predefined simulation box.
The argument \code{size} is of type string and denotes the major-axis length random generating
function name.

For the directional orientation of the cylinder axis one has the choice of a uniform
(\code{runifdir}), isotropic random planar (\code{rbetaiso}, see reference) or von Mises-Fisher
(\code{rvMisesFisher}) distribution. The simulation box is a list containing of vector arguments
which correspond to the lower and upper points in each direction. If the argument \code{box} has
only one element, i.e. \code{list(c(0,1)}, the same extent is used for the other dimensions.
If \code{rjoint="rmulti"} names a joint random generating function then argument \code{size} is ignored
(see example file "sim.R").
For the purpose of exact simulation [2] setting \code{size="rbinorm"} declares a bivariate normal
size-shape distribution. For a bivariate normal vector \eqn{[X,Y]} with correlation parameter \eqn{\rho}
the length of the cylinder is defined as \eqn{exp(x)} with (logit-transformed) shape parameter \eqn{s=1/(1+exp(-y))}.
This modification leads to a lognormally distributed length \code{h+2*r} of the cylinder where \code{h} is the height 
and \code{r} the radius of the (sphero)cylinder (and also the caps). The direction axis \code{u} is along the longer side of the
cylinder and independent of the size and shape. The following univariate distributions of the cylinder length and shape are also available:
`\code{rbeta}`, `\code{rgamma}`, `\code{rlnorm}` and `\code{runif}` distribution. Use `\code{const}` for a constant length or shape (radius) of the cylinders.
Only for `\code{rbinorm}` the exact simulation type can be used as an option if \code{perfect=TRUE}.
The current implementation does not include routines for unfolding the joint 3D size-shape-orientation
distribution of cylinders so far.

The argument \code{pl} denotes the print level of output information during simulation.
Currently, only \code{pl}=0 for no output and \code{pl}>100 for some additional info is implemented.
}
\examples{
\dontrun{

lam <- 10
box <- list("xrange"=c(0,3),"yrange"=c(0,3),"zrange"=c(0,9))

# cylinders of constant length
theta <- list("size"=list(0.25),
			  "shape"=list(0.5),
			  "orientation"=list("kappa"=1))

S <- simCylinderSystem(theta,lam,size="const", shape="const",
						orientation="rbetaiso",box=box,pl=101)
				
# cylinders of constant length with
# beta distributed radius	
theta <- list("size"=list(0.35),
              "shape"=list("a"=1,"b"=5),
			  "orientation"=list("kappa"=1.5))				
				
S <- simCylinderSystem(theta,lam,size="const", shape="rbeta",
				orientation="rbetaiso",box=box,pl=101)
				
# bivariate length-shape distribution
# possibly correlated 
param <- list("mx"=-1.0,"my"=-2.5, "sdx"=0.15,"sdy"=0.2,"rho"=0.0,"kappa"=1.0)
theta <- list("size"=list("mx"=param$mx,"sdx"=param$sdx,
						  "my"=param$my,"sdy"=param$sdy,
						  "rho"=param$rho),
			  "orientation"=list("kappa"=param$kappa),
			  "shape"=list())
		
S <- simCylinderSystem(theta,lam,size="rbinorm",orientation="rbetaiso",box=box,pl=101)	

## show cylinder system
#cols <- c("#0000FF","#00FF00","#FF0000","#FF00FF","#FFFF00","#00FFFF")
#cylinders3d(S, box, col=cols)
 

## TODO: joint with return (length,radius,theta,phi,u), shape is implicitly defined as r/length=s [0,1] 
	
}
}
\references{
\itemize{
	\item{} {Ohser, J. and Schladitz, K. 3D images of materials structures Wiley-VCH, 2009}
     \item{} {C. Lantu\eqn{\acute{\textrm{e}}}joul. Geostatistical simulation. Models and algorithms.
					Springer, Berlin, 2002. Zbl 0990.86007}
	  }
}
\author{
M. Baaske
}
