% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spheroid.R
\name{simSpheroidIntersection}
\alias{simSpheroidIntersection}
\title{Spheroid intersection}
\usage{
simSpheroidIntersection(theta, cond)
}
\arguments{
\item{theta}{simulation parameters}

\item{cond}{conditioning object for simulation and intersection}
}
\value{
list of intersection profiles
}
\description{
Simulate a spheroid system and intersect
}
\details{
The function first simulates a spheroid system according to the parameter \code{theta} and only
returns the resulting section profiles.
}
