% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sphere.R
\name{sphereIntersection}
\alias{sphereIntersection}
\title{Sphere intersection}
\usage{
sphereIntersection(S, d, n = c(0, 1, 0), intern = FALSE, pl = 0)
}
\arguments{
\item{S}{list of spheres, see \code{\link{simSphereSystem}}}

\item{d}{distance of the intersecting x-, y-plane to the origin}

\item{n}{normal vector of intersting plane}

\item{intern}{logical, \code{FALSE} (default), return all planar sections otherwise
only those which have their centers inside the intersecting window}

\item{pl}{print level, default \code{pl=0}}
}
\value{
either a list of section profiles or, if \code{pl=10}, a numeric vector of radii only
}
\description{
Intersect a sphere system
}
\details{
Given a sphere system obtained from \code{\link{simSphereSystem}}
the function returns the section radii from intersecting all spheres
stored in \code{S}.
}
\author{
M. Baaske
}
