% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spheroid.R
\name{spheroidIntersection}
\alias{spheroidIntersection}
\title{Intersect spheroid system}
\usage{
spheroidIntersection(S, d, n = c(0, 1, 0), intern = FALSE, pl = 0)
}
\arguments{
\item{S}{list of spheroids, see \code{\link{simSpheroidSystem}}}

\item{d}{distance of intersecting plane to the origin}

\item{n}{normal vector of intersting plane}

\item{intern}{logical, \code{FALSE} (default), return all section profiles otherwise
only those which have their centers inside the intersection window (if the 
intersected spheroid system had been simulated using exact simulation)}

\item{pl}{integer, \code{pl=0} (default), only return pointer to stored intersections}
}
\value{
list of size, shape and angle of section profiles
}
\description{
Intersect a spheroid system with a given plane
}
\details{
The function intersects a given spheroid system with a plane defined by a
normal vector, e.g. \code{n=c(0,1,0)}, perpendicular to one of the simulation bounding planes. 
The print level \code{pl} sets the type of return value. If \code{pl=10} the functions only returns
the lengths of the semiaxes, the shape factor of both and the angle in the intersection plane in \code{[0,pi/2]}.
}
\author{
M. Baaske
}
