% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/unfolding.R
\docType{data}
\name{spheroids}
\alias{spheroids}
\title{A packed spheroids configuration}
\format{List of spheroids where each list element
		   consists of the following variables:
\itemize{
	 \item{id: }{ current spheroid number}
	 \item{center: }{ x,y,z coordinates of the spheroid center}
	 \item{u: }{ orientation vector of spheroid's rotational symmetry axis}
	 \item{ab: }{ axes length}
	 \item{angles: }{ colatitude and longitude angle}
	 \item{rotM: }{ rotational matrix, center of rotation is the center of the spheroid}
}}
\description{
A dataset containing 12.606 non-overlapping spheroids
}
\keyword{datasets}

