% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spheroid.R
\name{spheroids3d}
\alias{spheroids3d}
\title{Plot particle system}
\usage{
spheroids3d(S, box, draw.axes = FALSE, draw.box = TRUE, clipping = FALSE,
  ...)
}
\arguments{
\item{S}{a list of spheroids}

\item{box}{simulation box}

\item{draw.axes}{logical: if true, draw the axes}

\item{draw.box}{logical: if true, draw the bounding box}

\item{clipping}{logical: if true clip to the bounding box}

\item{...}{further material properties passed to 3d plotting functions}
}
\description{
Draw particle system as defined by \code{S}.
}
\details{
The function requires the package \code{rgl} to be installed.
}

