% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unfolding.R
\name{trivarHist}
\alias{trivarHist}
\title{Trivariate histogram}
\usage{
trivarHist(A, main = paste("Trivariate Histogram"), scale = 0.5, col,
  ...)
}
\arguments{
\item{A}{3D array of count data (estimated histogram), see \code{\link{unfold}}}

\item{main}{main title of the plot}

\item{scale}{scaling factor for non-overlapping spheres}

\item{col}{vector of color values repeatedly used for all classes}

\item{...}{graphical parameters passed to \code{rgl::spheres3d}}
}
\description{
3D plot of a trivariate histogram
}
\details{
The (estimated spatial) joint size-shape-orientation distribution is plotted in a 3D
histogram with corresponding axes. The axes intersect in the first class number.
The ball volumes visualize the relative frequencies of count data for each class which
can be scaled by the user in order to make the spheres non-overlapping. Balls within the
same size class have the same color. For an example please see the file 'almmc.R'.
}
\author{
M. Baaske
}
