% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spheroid.R
\name{updateIntersections}
\alias{updateIntersections}
\title{Updated spheroid intersections}
\usage{
updateIntersections(S, id, cond)
}
\arguments{
\item{S}{spheroid system}

\item{id}{integer vector of spheroid id numbers}

\item{cond}{conditioning object}
}
\value{
integer vector indicating intersection=\code{1} or non intersection by \code{0}
}
\description{
Determine intersections of spheroids with bounding (simulation) box
}
\details{
For a given set of spheroid id numbers calculate if these spheroids intersect 
the simulation box or not.
}

