% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spheroid.R
\name{verticalSection}
\alias{verticalSection}
\title{Spheroid vertical section}
\usage{
verticalSection(S, d, n = c(1, 0, 0))
}
\arguments{
\item{S}{list of spheroids, see \code{\link{simSpheroidSystem}}}

\item{d}{distance of intersecting plane to the origin}

\item{n}{normal vector}
}
\value{
list of size, shape and angle of section profiles
}
\description{
Vertical section of spheroid system
}
\details{
The function performs a vertical intersection of the spheroid system either
parallel to yz-plane defined by the normal vector \code{n=c(1,0,0)} or
xz-plane defined by \code{n=c(0,1,0)}.
}

