% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unfolding.R
\name{binning3d}
\alias{binning3d}
\title{Histogram data}
\usage{
binning3d(size, angle, shape, breaks, check = TRUE, na.rm = TRUE)
}
\arguments{
\item{size}{vector of sizes}

\item{angle}{vector of angles}

\item{shape}{vector of aspect ratios}

\item{breaks}{list of bin vectors}

\item{check}{logical, default is \code{TRUE}}

\item{na.rm}{logical, if NAs are to be removed in the data vectors, default is \code{TRUE}}
}
\value{
3d array of count data
}
\description{
Count data of size, shape and orientation values.
}
\details{
For each value of planar or spatial measured quantities \code{size,shape,orientation}
the function counts the number of observations falling into each class.
The \code{breaks} list is either obtained from function \code{setbreaks} or the
user supplies his own variant, see \code{\link{setbreaks}} for details.
If \code{check} is \code{TRUE} some checks on the \code{breaks} list are done.
}
\author{
M. Baaske
}
