% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spheroid.R
\name{coefficientMatrixSpheroids}
\alias{coefficientMatrixSpheroids}
\title{Compute coefficients for trivariate unfolding}
\usage{
coefficientMatrixSpheroids(breaks, stype = c("prolate", "oblate"),
  check = TRUE, nCores = getOption("par.unfoldr", 1))
}
\arguments{
\item{breaks}{list of bin vectors}

\item{stype}{either \code{prolate} or \code{oblate}}

\item{check}{logical, whether to run some input checks}

\item{nCores}{number of cores used to compute the coefficients}
}
\value{
coefficient 6D array
}
\description{
Compute coefficients of the discretized integral equation for unfolding
}
\details{
In order to apply the EM algorithm to the stereological unfolding procedure for the joint
size-shape-orientation distribution in 3D one first has to compute the coefficients of the
discretized integral equation. This step is the most time consuming part of unfolding and
therefore has been separated in its own function. Further, the number of classes for size, shape and orientation
do not need to be the same, but the given class limits are also used for binning the spatial values. One can define
the number of cpu cores by a global option \code{par.unfoldr} or passing the number of cores by \code{nCores} directly.
}
\examples{
\dontrun{

#options(par.unfoldr=4)
breaks <- setbreaks(c(8,6,7),maxSize=0.37,kap=1.25)
breaks

P <- coefficientMatrixSpheroids(breaks,check=FALSE)
c(min(P),max(P),sum(P))

}
}
\author{
M. Baaske
}
