% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cylinder.R
\name{cylinderIntersection}
\alias{cylinderIntersection}
\title{Spherocylinder intersection}
\usage{
cylinderIntersection(S, d, n = c(0, 1, 0), intern = FALSE, pl = 0)
}
\arguments{
\item{S}{list of cylinders, see \code{\link{simCylinderSystem}}}

\item{d}{distance of intersecting plane to the origin}

\item{n}{normal vector of intersting plane}

\item{intern}{\code{intern=FALSE} (default) return all section profiles otherwise
only those which have their centers inside the intersection window}

\item{pl}{integer, \code{pl=0} (default), other options are not yet available}
}
\value{
list of size, shape and angle of section profiles
}
\description{
Intersect a spherocylinder system
}
\details{
The function intersects a cylinder system by an intersecting plane defined by a
normal vector \code{n=c(0,1,0)} (default), which depends on the main orientation
axis of the coordinate system.
}
\author{
M. Baaske
}
