% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spheroid.R
\name{cylinders3d}
\alias{cylinders3d}
\title{Cylinder system 3D}
\usage{
cylinders3d(S, box, draw.axes = FALSE, draw.box = TRUE, clipping = FALSE,
  ...)
}
\arguments{
\item{S}{list of cylinders, see \code{\link{simPoissonSystem}}}

\item{box}{simulation box}

\item{draw.axes}{logical, whether to show the axes}

\item{draw.box}{logical, whether to show the bounding box}

\item{clipping}{logical, whether to clip all lateral planes of the simulation box}

\item{...}{further material properties passed to 3d plotting functions}
}
\description{
Draw spherocylinders in 3D
}
\details{
The function requires the package \code{rgl} to be installed. For a full example please see
the file 'simCylinders.R'.
}
\author{
M. Baaske
}
