% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spheroid.R
\name{digitizeProfiles}
\alias{digitizeProfiles}
\title{Digitize section profiles}
\usage{
digitizeProfiles(sp, delta, win = NULL)
}
\arguments{
\item{sp}{list of section profiles, see \code{\link{intersectSystem}}}
}
\value{
image as a matrix
}
\description{
Digitize 2D objects (section profiles)
}
\details{
A list of section profiles, e.g. discs, ellipses or segments from intersected spherocylinders, is digitized according
to a given resolution such that the result is an image stored in matrix form.
}
\author{
M. Baaske
}
