% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spheroid.R
\name{drawSpheroidIntersection}
\alias{drawSpheroidIntersection}
\title{Plot Spheroid intersection}
\usage{
drawSpheroidIntersection(E, n = c(0, 1, 0), np = 25)
}
\arguments{
\item{E}{a list of spheroid intersections}

\item{n}{the normal vector of the intersecting plane}

\item{np}{number of points for polygon approximation of ellipses}
}
\description{
Drawing section profiles in 3D plane
}
\details{
The function requires the package \code{rgl} to be installed.
}
\author{
M. Baaske
}
