% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sphere.R
\name{em.saltykov}
\alias{em.saltykov}
\title{Expectation Maximization algorithm}
\usage{
em.saltykov(y, bin, maxIt = 32)
}
\arguments{
\item{y}{vector of observed absolute frequencies of circle diameters}

\item{bin}{non-decreasing vector of class limits}

\item{maxIt}{maximum number of iterations used}
}
\value{
vector of count data of absolute frequenties of sphere diameters
}
\description{
Estimation of empirical sphere diameter distribution
}
\details{
The function performs the EM algorithm.
}
\examples{
library(unfoldr)

## beta distribution for radii
lam <- 300
theta <- list("shape1"=2,"shape2"=4)
S <- simSphereSystem(theta,lam,rdist="rbeta",box=list(c(0,5)),
		## arguments only for returning intersection profiles
		dz=2.5,n=c(0,1,0),profiles=FALSE,intern=FALSE,		
		pl=100,label='N')

class(S)
length(S)

# planar vertical section
sp <- planarSection(S,d=2.5)
## or equivalently
## sp0 <- 2.0*sphereIntersection(S,2.5,c(0,0,1),FALSE,10)
## sp[1:10]
## sp0[1:10]

ret <- unfold(sp,nclass=25)
 
## Point process intensity
cat("Intensities: ", sum(ret$N_V)/25, "vs.",lam,"\\n")
 
## original diameters
r3d <- unlist(lapply(S,function(x) 2.0*x$r))
rest3d <- unlist(lapply(2:(length(ret$breaks)),
            function(i) rep(ret$breaks[i],sum(ret$N_V[i-1]))))
 
op <- par(mfrow = c(1, 2))
hist(r3d[r3d<=max(ret$breaks)], breaks=ret$breaks, main="Radius 3d",
     freq=FALSE, col="gray",xlab="r")
hist(rest3d, breaks=ret$breaks,main="Radius estimated",
     freq=FALSE, col="gray", xlab="r")
par(op)
}
\references{
Ohser, J. and Muecklich, F. Statistical analysis of microstructures in materials science J. Wiley & Sons, 2000
}
\author{
M. Baaske
}
