% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unfolding.R
\name{parameters3d}
\alias{parameters3d}
\title{3D characteristics of spheroids}
\usage{
parameters3d(S)
}
\arguments{
\item{S}{list of spheroids}
}
\value{
a list of spheroids` 3D characteristics
}
\description{
Extract size, shape and orientation angle
}
\details{
The function extracts the characteristics of a 3D spheroid system, either oblate or prolate spheroids,
and returns a list of the following elements: major semi-axis length '\code{a}', shape factor '\code{s}'
and polar angle '\code{Theta}'. For a full example please see the file 'unfold.R'.
}
\author{
M. Baaske
}
