% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spheroid.R
\name{planarSection}
\alias{planarSection}
\title{Sphere planar section}
\usage{
planarSection(S, d, intern = FALSE, pl = 0)
}
\arguments{
\item{S}{list of spheres of class \code{sphere}, see \code{\link{simPoissonSystem}}}

\item{d}{distance of the intersecting xy-plane to the origin (as a planar section)}

\item{intern}{logical, \code{FALSE} (default), return all planar section profiles otherwise
only those which have their centers inside the intersecting window}

\item{pl}{print level, \code{pl=0} (default) for no output}
}
\value{
numeric vector of planar section diameters
}
\description{
Intersection of given spheres by a plane
}
\details{
The function intersects a sphere system and returns the diameters of the corrsponding planar section profiles.
}
\author{
M. Baaske
}
