% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spheroid.R
\name{planarSection}
\alias{planarSection}
\title{Sphere planar section}
\usage{
planarSection(S, d, intern = FALSE, pl = 0)
}
\arguments{
\item{S}{list of spheres of class \code{sphere}, see \code{\link{simPoissonSystem}}}

\item{d}{distance of the (planar) xy-plane to the origin}

\item{intern}{logical, \code{FALSE} (default), whether to return all section profiles or
only those which have their centers inside the intersecting window}

\item{pl}{print level, \code{pl>0} for verbose output}
}
\value{
numeric vector of planar section diameters
}
\description{
Planar intersection of spheres
}
\details{
The function computes the planar intersection of a sphere system and returns the diameters
of the corrsponding section profiles.
}
\author{
M. Baaske
}
