% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spheroid.R
\name{setupSpheroidSystem}
\alias{setupSpheroidSystem}
\title{Setup spheroid system}
\usage{
setupSpheroidSystem(S, box = list(c(0, 1)), mu = c(0, 0, 1),
  perfect = TRUE, pl = 0, label = "N")
}
\arguments{
\item{S}{spheroid system}

\item{box}{simulation box}

\item{mu}{main orientation axis, \code{mu=c(0,0,1)} (default)}

\item{perfect}{logical: \code{perfect=TRUE} (default) simulate perfect}

\item{pl}{integer, \code{pl=0} (default), print level}

\item{label}{character, \code{"N"} (default), set a label to the generated spheroids}
}
\value{
pointer to C storage structure which holds the spheroid system \code{S}
}
\description{
Reinitialize spheroid system after R workspace reloading
}
\details{
The internal data structure has to be reinitialized if a spheroid system was reloaded to the R workspace.
Calling this function is needed only in case one desires to get profile sections of the spheroid
system again which has been previously stored as an R (list) object and afterwards
reloaded.
}
\author{
M. Baaske
}
