% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spheroid.R
\name{setupSpheroidSystem}
\alias{setupSpheroidSystem}
\title{Setup spheroid system}
\usage{
setupSpheroidSystem(S, mu = c(0, 1, 0), pl = 0)
}
\arguments{
\item{S}{spheroid system}

\item{mu}{main orientation vector}

\item{pl}{print level}
}
\value{
\code{NULL}
}
\description{
Reinitialize spheroid system after R workspace reloading
}
\details{
The internal data structure has to be reinitialized if a spheroid system was reloaded to the R workspace.
Calling this function is needed only in case one desires to get profile sections of the spheroid
system again which has been previously stored as an R (list) object and afterwards
reloaded.
}
\author{
M. Baaske
}
