% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spheroid.R
\name{simPoissonSystem}
\alias{simPoissonSystem}
\title{Poisson germ-grain process}
\usage{
simPoissonSystem(theta, lam, size = "const", shape = "const",
  orientation = "rbetaiso", type = c("prolate", "oblate", "spheres",
  "cylinders"), rjoint = NULL, box = list(c(0, 1)), mu = c(0, 0, 1),
  dz = 0, n = c(0, 1, 0), intersect = c("original", "full", "only",
  "digi"), delta = 0.01, intern = FALSE, perfect = FALSE, pl = 0,
  label = "N")
}
\arguments{
\item{theta}{list of simulation parameters which must consist of elements: \code{size}, \code{shape} and \code{orientation}}

\item{lam}{mean number of objects per unit volume}

\item{size}{name of the size distribution function}

\item{shape}{name of the shape distribution function}

\item{orientation}{name of direction distribution function}

\item{type}{spheroid type, either "\code{prolate}" or "\code{oblate}" can be used}

\item{rjoint}{user-defined function, which specifies the (joint) distribution of the size, shape and orientation}

\item{box}{simulation box}

\item{mu}{main orientation axis, \code{mu=c(0,0,1)} (default), as the alignment of the coordinate system}

\item{dz}{distance of the intersecting plane to the origin}

\item{n}{normal vector of the intersting plane}

\item{intersect}{set "\code{full}" to return the simulated system together with section profiles as two lists named \code{S} and \code{sp} respectively,
choose "\code{only}" for section profiles only,
"\code{original}" for the 3D system only and "\code{digi}" for an integer matrix \code{W} as an discretized (binary) image
of section profiles}

\item{delta}{lattice constant for discretization, set to \code{0.01} (default)}

\item{intern}{logical, \code{FALSE} (default), whether to return only section profiles with centers inside the simulation window}

\item{perfect}{logical, \code{FALSE} (default), whether to simulate exactly (also called perfect)}

\item{pl}{integer, print level and return value type, see details}

\item{label}{character, a label set to each generated object, set to '\code{N}' (default)}
}
\value{
list of 3D objects depending on the chosen return type defined by the argument \code{intersect}
}
\description{
Simulation of a Poisson germ-grain process with either spheres, spheroids or spherocylinders as grains
}
\details{
The function simulates a Poisson germ-grain process given certain form parameters in the argument \code{theta}
and a predefined simulation box. The positions of the germs are generated independently according to a Poisson
process with mean intensity parameter \code{lam}. The function either generates \code{type="prolate"} or \code{type="oblate"}
spheroids, spheres or spherocylinders. The argument \code{size} sets the name of the distribution function of the size
of the objects, i.e. the semi-major axis lengths in case of spheroids, radii for spheres or the lengths of the main axis
of rotation for spherocylinders including the caps.

The following directional (orientation) distributions of the spheroid's major-axis or cylinder's main axis are available:
a uniform distribution (\code{runifdir}), isotropic random planar distribution (\code{rbetaiso}, see the reference below)
and a "von Mises-Fisher" (\code{rvMisesFisher}) distribution. The simulation box is a list containing the ranges of each box dimension
corresponding to the lower and upper limits in each direction. If the argument \code{box} contains only a single range, i.e. \code{list(c(0,1)}, the
same limits are assumed for the remaining dimensions. The argument \code{rjoint} defines a (joint) distribution function which can be any function
provided by the user in order to generate the required random parameters for spheroids or spherocylinders. An in-depth example of usage is given in
the example files 'simSpheroids.R' and 'simCylinders.R'.

In addition, the function supports an exact simulation type [2] of the grains. In case of spheroids and spherocylinders setting \code{size="rbinorm"}
declares a bivariate normal size-shape distribution for which the exact simulation is available setting \code{perfect=TRUE}. More specifically,
for a bivariate normal vector \eqn{[X,Y]} with correlation parameter \eqn{\rho}, the length of the semi-major axis of a spheroid is given by \eqn{a=exp(x)}
with a (logit-transformed) shape parameter as \eqn{s=1/(1+exp(-y))} and thus a scaled semi-minor axis length \eqn{c=a*s}. This modification leads to a lognormally
distributed length of the semi-major axis. In case of spherocylinders, the lognormally distributed length of a cylinder is \eqn{len=h+2*r} where
\code{h} is the height and \eqn{r=len/2*s} the radius. The main direction \code{u} of the spheroid or spherocylinder is determined by the
major axis independent of size and shape. Also, the following univariate distributions of the length \code{a} and, respectively, \code{len},
the shape \code{s} are available: `\code{rbeta}`, `\code{rgamma}`, `\code{rlnorm}` and `\code{runif}`. One can also use "\code{const}" for a
simulation with constant lengths or shapes. Only simulations with size distributions `\code{rbinorm}` or `\code{rlnorm}` can use the exact
type of simulation (\code{perfect=TRUE}).

For spheres any distribution of the radii can be specified as a name of a user-defined function in \code{size} as long as the formal named
function parameters match the actual named parameters exactly as defined in the parameters given by \code{theta}.
Besides this, all other distributions given above are also available. In particular, setting \code{size="rlnorm"} leads to lognormally distributed
radii of the spheres in which case the exact simulation is available as an option (\code{perfect=TRUE}). Use "\code{const}" for a constant
radius of simulated spheres. 
 
The argument \code{pl>=0} denotes both the print level of intermediate output and the type of return value. If \code{pl=10}, then a shorter list
of spheroids or spheres is returned to speed up computation. Note that, the current implementation does not include routines for unfolding
the joint 3D size-shape-orientation distribution of spherocylinders so far.
}
\examples{
 # intensity parameter
	lam <- 100

 # simulation bounding box
 box <- list("xrange"=c(0,5),"yrange"=c(0,5),"zrange"=c(0,5))

 # log normal size distribution with a constant shape factor and
 # isotropic (\\code{kappa=1}) random orientation, see reference [1] 
 theta <- list("size"=list("meanlog"=-2.5,"sdlog"=0.5),
               "shape"=list("s"=0.5),
               "orientation"=list("kappa"=1))

 S <- simPoissonSystem(theta,lam,size="rlnorm",box=box,type="oblate",pl=1) 
length(S)

}
\references{
\itemize{
	\item{} {Ohser, J. and Schladitz, K. 3D images of materials structures Wiley-VCH, 2009}
     \item{} {C. Lantu\eqn{\acute{\textrm{e}}}joul. Geostatistical simulation. Models and algorithms.
					Springer, Berlin, 2002. Zbl 0990.86007}
	  }
}
\author{
M. Baaske
}
