% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sphere.R
\name{simSphereSystem}
\alias{simSphereSystem}
\title{Simulation of sphere system}
\usage{
simSphereSystem(theta, lam, rdist, box = list(c(0, 1)), dz = 0, n = c(0,
  1, 0), profiles = FALSE, intern = FALSE, perfect = TRUE, pl = 0,
  label = "N")
}
\arguments{
\item{theta}{simulation parameters}

\item{lam}{mean number of spheres per unit volume}

\item{rdist}{name of radii distribution}

\item{box}{simualtion box}

\item{dz}{distance of the intersecting x-, y-plane to the origin}

\item{n}{normal vector of intersting plane}

\item{profiles}{logical, \code{profiles=FALSE} (default), whether the simulated system of spheres is intersected afterwards in which case
only sections profiles are returned}

\item{intern}{logical, \code{intern=FALSE} (default), whether to return only section profiles with centers inside the simulation window}

\item{perfect}{logical, \code{perfect=TRUE} (default), exact simulation}

\item{pl}{integer, print level and return value definition}

\item{label}{character, label passed to each simulated sphere, ´\code{N}´ (default)}
}
\value{
The function either returns a list of spheres with elements \code{id}, \code{center} and radius \code{r} of class \code{spheres}
 	   or a vector of radii.
}
\description{
The function simulates a Poisson sphere system.
}
\details{
Any distribution for the radii can be specified as a name in `\code{rdist}`as long as the formal named function parameters
match the actual named parameters exactly as defined in the parameter list `\code{theta}`.

The simulation box is given by a list of length three. Each element is a vector corresponding to the lower and upper points in x-, y- and z-direction in
the sense of a canonically oriented simulation box. If \code{box} has only one element, i.e. \code{list(c(0,1)}, the same extent
is used for the remaining dimensions. The argument \code{pl} denotes the print level of information during simulation and is used to set the type of the
return value. If \code{pl=10} only the radii are returned otherwise a full list of spheres. The argument `\code{rdist}` declares
the name of the (user defined) radii distribution. Setting \code{rdist="rlnorm"} leads to lognormally distributed
radii of the spheres. In this case an exact simulation [1] is available if \code{perfect=TRUE}. Currently available other
distributions are defined by `\code{rbeta}`, `\code{rgamma}`, `\code{rlnorm}` and `\code{runif}`. Use `\code{const}` for a constant
radius of all spheres.
}
\examples{
 theta <- list("meanlog"=-2.5,"sdlog"=0.2)
 S <- simSphereSystem(theta,lam=1000,rdist="rlnorm",pl=101)

}
\references{
\itemize{
   \item{} {C. Lantu\eqn{\acute{\textrm{e}}}joul. Geostatistical simulation. Models and algorithms.
            Springer, Berlin, 2002. Zbl 0990.86007}
	 }
}
\author{
M. Baaske
}
