% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spheroid.R
\name{spheroids3d}
\alias{spheroids3d}
\title{Spheroid system 3D}
\usage{
spheroids3d(S, box, draw.axes = FALSE, draw.box = TRUE, draw.bg = TRUE,
  bg.col = "white", clipping = FALSE, ...)
}
\arguments{
\item{S}{list of spheroids, see \code{\link{simPoissonSystem}}}

\item{box}{simulation box}

\item{draw.axes}{logical, whether to show the axes}

\item{draw.box}{logical, whether to show the bounding box}

\item{draw.bg}{logical, whether to show a gray background box correpsonding to the simulation box}

\item{bg.col}{background color used showw the box background}

\item{clipping}{logical, whether to clip all lateral planes of the simulation box}

\item{...}{further material properties passed to 3D plotting functions}
}
\description{
Draw a spheroid system in 3D
}
\details{
The function requires the package '\code{rgl}' to be installed. For a full example
please see the file 'simSpheroids.R'.
}
\author{
M. Baaske
}
