% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spheroid.R
\name{spheroids3d}
\alias{spheroids3d}
\title{Draw a spheroid system in 3D}
\usage{
spheroids3d(S, box, draw.axes = FALSE, draw.box = TRUE, draw.bg = TRUE,
  bg.col = "white", clipping = FALSE, ...)
}
\arguments{
\item{S}{list of spheroids, see \code{\link{simPoissonSystem}}}

\item{box}{simulation box}

\item{draw.axes}{logical: if true, draw the axes}

\item{draw.box}{logical: if true, draw the bounding box}

\item{draw.bg}{logical: if true, draw the a gray background box}

\item{bg.col}{background color used to draw the box background}

\item{clipping}{logical: if true clip to the bounding box}

\item{...}{further material properties passed to 3d plotting functions}
}
\description{
Draw a spheroid system in 3D
}
\details{
The function requires the package \code{rgl} to be installed.
}
\author{
M. Baaske
}
