% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unfolding.R
\name{trivarHist}
\alias{trivarHist}
\title{Trivariate histogram}
\usage{
trivarHist(A, main = paste("Trivariate Histogram"), scale = 0.5, col,
  ...)
}
\arguments{
\item{A}{3d array of count data}

\item{main}{main title of the plot}

\item{scale}{factor to scale the spheres}

\item{col}{vector of color values repeatedly used}

\item{...}{optional argumentspassed to function \code{rgl::spheres3d}}
}
\description{
Plot trivariate histogram of joint size-shape-orientation distribution
}
\details{
The (estimated spatial) joint size-shape-orientation distribution is plotted in a 3D
histogram with corresponding axesn. The axes intersect in the first class number.
The ball volumes visualize the relative frequencies of count data for each class which
can be scaled by the user for non-overlapping spheres. Balls within the same size class
have the same color. For an example please see the file 'almmc.R'.
}
\author{
M. Baaske
}
