% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spheroid.R
\name{updateIntersections}
\alias{updateIntersections}
\title{Check intersection}
\usage{
updateIntersections(S)
}
\arguments{
\item{S}{list of spheres, spheroids or spherocylinders, see \code{\link{simPoissonSystem}}}
}
\value{
binary integer vector of length equal to the length of \code{S}.
}
\description{
Check itersection of either spheres, spheroids or spherocylinders with the simulation box
}
\details{
For a given list of spheres, spheroids or spherocylinders the function tests whether an object intersects
one of the lateral planes (bottom, top plane included) of the simulation box. The vector returned is of length
equal to the number of objects in \code{S} and has entries either \code{1} for an object which is intersected by
and \code{0} otherwise.
}
\author{
M. Baaske
}
