% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spheroid.R
\name{updateIntersections}
\alias{updateIntersections}
\title{Spheroid intersections with box}
\usage{
updateIntersections(S, box)
}
\arguments{
\item{S}{geometric objects system}

\item{box}{the simulation box}
}
\value{
integer vector indicating intersection=\code{1} or non intersection by \code{0}
}
\description{
Determine intersections of spheroids with bounding (simulation) box
}
\details{
For a given list of spheroids, spheres or cylinders the function tests whether the objects intersect
the simulation bounding box.
}
\author{
M. Baaske
}
