% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spheroid.R
\name{verticalSection}
\alias{verticalSection}
\title{Spheroid vertical section}
\usage{
verticalSection(S, d, n = c(0, 1, 0), intern = FALSE)
}
\arguments{
\item{S}{list of spheroids, see \code{\link{simSpheroidSystem}}}

\item{d}{distance of intersecting plane to the origin}

\item{n}{normal vector of intersting plane}

\item{intern}{logical, \code{FALSE} (default), return all section profiles otherwise
only those which have their centers inside the intersection window (if the 
intersected spheroid system had been simulated using exact simulation)}
}
\value{
list sizes \code{A}, shapes \code{S} and angles \code{alpha} of section profiles
}
\description{
Vertical section of a spheroid system
}
\details{
The function intersects a spheroid system by an intersecting plane defined by a normal 
vector \code{n=c(0,1,0)} (default), which depends on the main orientation axis of the
coordinate system.
}
\author{
M. Baaske
}
